+++ app/src/main/java/com/monkeyteam/chimpagne/ui/navigation/NavigationActions.kt
@@ -10,11 +10,9 @@ object Route {
   const val ACCOUNT_EDIT_SCREEN = "AccountEdit"
   const val HOME_SCREEN = "Home"
   const val FIND_AN_EVENT_SCREEN = "FindAnEvent"
-  const val LOADING = "loading"
   const val EVENT_CREATION_SCREEN = "EVENT_CREATION_SCREEN"
   const val MY_EVENTS_SCREEN = "myEvents"
   const val EVENT_SCREEN = "viewDetailEventScreen"
-  const val ONLINE_EVENT_VIEW = "onlineEventView"
   const val LOADING_LOGIN = "loadingLogin"
   const val MANAGE_STAFF_SCREEN = "manageStaffScreen"
   const val SUPPLIES_SCREEN = "SuppliesScreen"
